Draft: profile view item (#188)
 style: profile view item
adding an alert before deleting an item
toast added
closes: #187

closes: #188

+++ src/components/ProfileContentCard/ProfileContentCard.jsx
@@ -1,12 +1,20 @@
 import React from 'react'
-
+import { useDispatch } from 'react-redux'
 import moment from 'moment'
+import { useHistory } from 'react-router-dom'
 import { useTranslation } from 'react-i18next'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import PropTypes from 'prop-types'
+import { useToasts } from 'react-toast-notifications'
 import { db } from '../../firebase'
+import { FetchProducts } from '../../redux'
 
 export default function ProfileContentCard({ title, content, seen, time, id }) {
+  const { addToast } = useToasts()
+
+  const history = useHistory()
+  const dispatch = useDispatch()
+
   const { t } = useTranslation()
   function timing(date) {
     const now = moment()
@@ -20,10 +28,16 @@ export default function ProfileContentCard({ title, content, seen, time, id }) {
   }
 
   const deleteItem = productId => {
-    db.collection('products')
-      .doc(productId)
-      .delete()
-      .then(() => window.location.reload(false))
+    if (window.confirm(t('profile.deleting'))) {
+      db.collection('products')
+        .doc(productId)
+        .delete()
+        .then(dispatch(FetchProducts()))
+      addToast('Item Deleted', { appearance: 'warning' })
+    }
+  }
+  const handleClick = () => {
+    history.push(`/productdetails/${id}`)
   }
 
   return (
@@ -47,10 +61,19 @@ export default function ProfileContentCard({ title, content, seen, time, id }) {
         <div className="justify-self-end self-center ">
           <button
             type="button"
-            className="bg-red px-4 py-1 rounded-full text-white hover:shadow-none focus:outline-none shadow-md hover:bg-darkRed transition duration-300 ease-in-out "
-            onClick={() => deleteItem(id)}
+            className="bg-yellow px-4 py-1 rounded-full text-pureWhite hover:shadow-none focus:outline-none shadow-md  transition duration-300 ease-in-out "
+            onClick={handleClick}
           >
-            {t('profile.delete')}
+            <FontAwesomeIcon icon="binoculars" />
+          </button>
+          <button
+            type="button"
+            className="bg-red mx-2 px-4 py-1 rounded-full text-white hover:shadow-none focus:outline-none shadow-md  transition duration-300 ease-in-out "
+            onClick={() => {
+              deleteItem(id)
+            }}
+          >
+            <FontAwesomeIcon icon="trash-alt" />
           </button>
         </div>
       </div>
